pyproject.toml
[project]
name = "xerxes-agent"
version = "0.2.0"
description = "Agents for intelligence and coordination"
authors = [{ name = "Erfan Zare Chavoshi", email = "Erfanzare810@gmail.com" }]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.11,<3.14"
keywords = ["xerxes", "Agents", "AI", "LLM", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",

    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

# Core runtime dependencies
dependencies = [
    "pydantic>=2.9.2,<3.0.0",
    "openai>=1.72.0",
    "httpx>=0.25.0",
    "psutil>=5.9.0",
    "pillow>=11.0.0",
    "numpy>=1.24.0",
    "google-generativeai>=0.8.0,<0.9.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "jinja2>=3.1.6",
    "requests>=2.31.0",
    "ddgs>=9.5.4",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
    "lxml>=5.0.0",
    "playwright>=1.54.0",
]

[project.urls]
Homepage = "https://github.com/erfanzar/Xerxes-Agents"
Repository = "https://github.com/erfanzar/Xerxes-Agents"
Documentation = "https://erfanzar.github.io/Xerxes-Agents"

[project.scripts]
xerxes = "xerxes.__main__:main"

[project.optional-dependencies]
# Development tools
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.7.0",
    "black>=24.8.0",
    "mypy>=1.12.0",
    "pre-commit>=3.6.0",
    "ipykernel>=6.29.0",
    "notebook>=7.0.0",
]

ui = ["chainlit~=2.9.2", "markdown>=3.5.0"]

# Search and web tools
search = []

# Database backends (currently unused - placeholder for future)
# database = [
#     "sqlalchemy>=2.0.0",
#     "alembic>=1.13.0",
#     "asyncpg>=0.29.0",
#     "psycopg2-binary>=2.9.9",
#     "pymongo>=4.6.0",
#     "motor>=3.3.0",
# ]

# Redis for caching/queuing (currently unused - placeholder for future)
# redis = ["redis>=5.0.0", "redis[hiredis]>=5.0.0"]

# Monitoring and observability
monitoring = [
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation>=0.43b0",
    "sentry-sdk>=1.40.0",
    "datadog>=0.48.0",
    "psutil>=5.9.0",
]

# Vector search and embeddings
vectors = [
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.3.0",
    # "faiss-cpu>=1.7.4",  # Currently unused - uncomment if needed
]

# MCP (Model Context Protocol) - for SSE and Streamable HTTP transports
mcp = [
    "mcp>=1.1.0",
]

# Alternative LLM providers (Note: these use httpx which is already in core)
# llm-providers = [
#     "anthropic>=0.39.0",  # Not actually used - AnthropicLLM uses httpx directly
#     "ollama>=0.3.0",      # Not actually used - OllamaLLM uses httpx directly
# ]

# Full installation with all features
full = ["xerxes-agent[search,monitoring,vectors,ui,mcp]"]

# Common presets for different use cases
research = ["xerxes-agent[search,vectors,ui]"]
enterprise = ["xerxes-agent[monitoring]"]
ai-complete = ["xerxes-agent[vectors]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/python/xerxes"]

[tool.hatch.build.targets.wheel.force-include]
"src/python/xerxes/py.typed" = "xerxes/py.typed"


[tool.black]
line-length = 121
target-version = ["py311"]
preview = true

[tool.ruff]
line-length = 121
target-version = "py311"

[tool.ruff.lint]
select = ["A", "B", "E", "F", "I", "NPY", "RUF", "UP", "W"]
ignore = ["F722", "B008", "UP015", "A005", "E501", "B010"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
exclude = ["tests/", "examples/", "scripts/", "docs/"]

[tool.basedpyright]
# Disabled diagnostic rules to reduce type-checking noise
reportAssignmentType = "none"
reportUnknownVariableType = "none"
reportPrivateUsage = "none"
reportOptionalCall = "none"
reportUnknownMemberType = "none"
reportGeneralTypeIssues = "none"
reportArgumentType = "none"
enableTypeIgnoreComments = true
reportMissingTypeArgument = "none"
reportUnknownParameterType = "none"
reportAny = "none"
reportCallIssue = "none"
reportUnknownArgumentType = "none"
reportRedeclaration = "none"
reportIgnoreCommentWithoutRule = "none"
reportUnnecessaryIsInstance = "none"
reportUnusedCallResult = "none"
reportUnknownLambdaType = "none"
reportUnannotatedClassAttribute = "none"
reportImplicitStringConcatenation = "none"
reportMissingParameterType = "none"
reportImplicitOverride = "none"
reportImplicitAbstractClass = "none"
reportExplicitAny = "none"
reportIncompatibleMethodOverride = "none"
reportIndexIssue = "none"
reportAttributeAccessIssue = "none"
reportUnreachable = "none"
reportOptionalMemberAccess = "none"
reportOperatorIssue = "none"
reportUninitializedInstanceVariable = "none"
reportIncompatibleVariableOverride = "none"

[tool.uv.build-backend]
module-name = "xerxes-agent"
module-root = "src/python"

[tool.setuptools]
package-dir = { "" = "src/python" }

[tool.setuptools.packages.find]
where = ["src/python"]
